dom/devicestorage/test/test_lastModificationFilter.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup();

var oldFiles = ["a.png", "b.png", "c.png"];
var timeFile = "t.png";
var newFiles = ["d.png", "e.png", "f.png"];

var storage = navigator.getDeviceStorage('pictures');
var prefix = "devicestorage/" + randomFilename(12);
var i;
var timestamp;
var lastFileAddedTimestamp;

function verifyAndDelete(prefix, files, e) {
  if (e.target.result == null) {
    ok(files.length == 0, "when the enumeration is done, we shouldn't have any files in this array")
    dump("We still have length = " + files.length + "\n");
    dump(files + "\n");
    devicestorage_cleanup();
    return;
  }

  var filename = e.target.result.name;
  if (filename[0] == "/") {
    // We got /storageName/prefix/filename
    // Remove the storageName (this shows up on FirefoxOS)
    filename = filename.substring(1); // Remove leading slash
    var slashIndex = filename.indexOf("/");
    if (slashIndex >= 0) {
      filename = filename.substring(slashIndex + 1); // Remove storageName
    }
  }
  if (filename.startsWith(prefix)) {
    filename = filename.substring(prefix.length + 1); // Remove prefix
  }

  var index = files.indexOf(filename);
  ok(index > -1, "filename should be in the enumeration : " + e.target.result.name);
  if (index == -1)
    return;

  files.splice(index, 1);

  // clean up
  var cleanup = storage.delete(e.target.result.name);
  cleanup.onsuccess = function(e) {}
}

function addFile(filename, callback) {
  var addReq = storage.addNamed(createRandomBlob('image/png'), prefix + '/' + filename);
  addReq.onsuccess = function(e) {
    // After adding the file, we go ahead and grab the timestamp of the file
    // that we just added
    var getReq = storage.get(prefix + '/' + filename);
    getReq.onsuccess = function(e) {
      lastFileAddedTimestamp = e.target.result.lastModifiedDate;
      callback();
    }
    getReq.onerror = function(e) {
      ok(false, "getError was called : " + e.target.error.name);
      devicestorage_cleanup();
    };
  }
  addReq.onerror = function(e) {
    ok(false, "addError was called : " + e.target.error.name);
    devicestorage_cleanup();
  }
}

function addFileArray(fileArray, callback) {
  var i = 0;
  function addNextFile() {
    i = i + 1;
    if (i == fileArray.length) {
      callback();
      return;
    }
    addFile(fileArray[i], addNextFile);
  }
  addFile(fileArray[0], addNextFile);
}

function delFile(filename, callback) {
  var req = storage.delete(prefix + '/' + filename);
  req.onsuccess = function(e) {
    callback();
  };
  req.onerror = function(e) {
    ok(false, "delError was called : " + e.target.error.name);
    devicestorage_cleanup();
  };
}

function afterNewFiles() {
  var cursor = storage.enumerate(prefix, {"since": timestamp});
  cursor.onsuccess = function(e) {
    verifyAndDelete(prefix, newFiles, e);
    if (e.target.result) {
      e.target.continue();
    }
  };
  cursor.onerror = function (e) {
    ok(false, "handleError was called : " + e.target.error.name);
    devicestorage_cleanup();
  };
}

function waitForTimestampChange() {
  // We've added a new file. See if the timestamp differs from
  // oldFileAddedTimestamp, and if it's the same wait for a bit
  // and try again.
  if (lastFileAddedTimestamp.valueOf() === timestamp.valueOf()) {
    delFile(timeFile, function() {
      setTimeout(function() {
        addFile(timeFile, waitForTimestampChange);
      }, 1000);
    });
  } else {
    timestamp = lastFileAddedTimestamp;
    // The timestamp has changed. Go ahead and add the rest of the new files
    delFile(timeFile, function() {
      addFileArray(newFiles, afterNewFiles);
    });
  }
}

function afterOldFiles() {
  timestamp = lastFileAddedTimestamp;
  setTimeout(function() {
    // We've added our old files and waited for a second.
    // Add a new file until the timestamp changes
    addFile(timeFile, waitForTimestampChange);
  }, 1000);
}

function addOldFiles() {
  addFileArray(oldFiles, afterOldFiles);
}

addOldFiles();

</script>
</pre>
</body>
</html>
